modified:   set0/einstein.py

+++ set0/einstein.py
@@ -1,2 +1,2 @@
-# Assume that the user will always input an int or a floating point
-print(float(input('Enter the mass: '))*(9*10**16))
+# Assume that the user will always input an int
+print(int(input('Enter the mass: '))*(9*(10**16)))
